+++ internal/generator/internal/scope/file.go
@@ -13,6 +13,7 @@ type File struct {
 	GoPackageName string
 
 	messages []*Message
+	enums    []*Enum
 }
 
 // Messages returns the set of messages defined in this file.
@@ -30,4 +31,21 @@ func (f *File) Messages() []*Message {
 		}
 	}
 	return f.messages
+}
+
+// Enums returns the set of enum types defined in this file.
+func (f *File) Enums() []*Enum {
+	if f.enums == nil {
+		for _, d := range f.Descriptor.GetEnumType() {
+			f.enums = append(
+				f.enums,
+				&Enum{
+					File:       f,
+					Descriptor: d,
+					GoTypeName: identifier.Exported(d.GetName()),
+				},
+			)
+		}
+	}
+	return f.enums
 }